Caching graphs politics set for 5 minutes. Kotlin upgrade. Model update.

+++ src/main/java/com/coinpaprika/apiclient/entity/QuoteEntity.kt
@@ -15,7 +15,7 @@ import kotlinx.android.parcel.Parcelize
 data class QuoteEntity(@SerializedName("price") val price: Double,
                        @SerializedName("volume_24h") val dailyVolume: Double,
                        @SerializedName("volume_24h_change_24h") val dailyVolumeDailyChange: Double,
-                       @SerializedName("market_cap") val marketCap: Long,
+                       @SerializedName("market_cap") val marketCap: Double,
                        @SerializedName("market_cap_change_24h") val marketCapDailyChange: Double,
                        @SerializedName("percent_change_1h") val percentChange1h: Double,
                        @SerializedName("percent_change_12h") val percentChange12h: Double,
